swarm/pss: split pss.go file
